Allow either `kube-linter` or `tee` command fail the action
Previously we returned only `${PIPESTATUS[0]}` but that masked error
from `tee` command when, for example, it could not write file.

See https://github.com/stackrox/kube-linter-action/runs/3743081013#step:3:45
where the step is still listed as successful despite that tee failed
to write file to the directory `sarif-results/` that does not exist.